requirementsController.txt
Django>=3.0.7
djangorestframework==3.11.0
psycopg2-binary==2.8.3
requests==2.22.0
sh==1.12.14
